feat(help): swap custom ANSI renderer for charmbracelet/glamour
Glamour is a production-grade markdown-to-terminal renderer used by
gh, glow, and many other Charm CLIs. Benefits over the custom
~200-LOC ANSI renderer it replaces:
- Syntax-highlighted fenced code blocks (via chroma)
- Correct word wrap at column 80
- Richer inline styling (links, distinct italic, tighter spacing)
- Consistent dark theme for TTY; stripped plain text for non-TTY

The custom Tokenize() + Token types are deleted; they were only
used by the old text renderer. FindUnsupported (the content-
discipline linter) stays, moved to linter.go — it enforces what
authors can write, regardless of how it renders.

API deviation from the plan sketch: WithAutoStyle() uses os.Stdout.Fd()
for TTY detection (not the injected writer), so it emits notty style
in tests. Switched to explicit WithStandardStyle("dark") for isTTY=true
— same visual quality, deterministic in tests.

Size: +~6.5 MB binary (glamour + lipgloss + chroma), accepted for
output quality.
